+++ docs/history.rst
@@ -4,6 +4,8 @@ Version History
 :compare:`Next <3.0.4...master>`
 --------------------------------
 - Add type annotations (see :ref:`type-info`)
+- Return a "406 Not Acceptable" if the :http:header:`Accept` header values cannot be matched
+  and there is no default content type configured
 
 :compare:`3.0.4 <3.0.3...3.0.4>` (2 Nov 2020)
 ---------------------------------------------